Relocate mock_galaxy from tests to core for cli test entrypoint
